requirements.txt
Django==5.2
Pillow>=9.0.0
python-dotenv>=0.19.0
channels>=3.0.0
requests>=2.20.0
